# docs/swagger.yaml
basePath: /api
definitions:
...
  models.Transaction:
    properties:
      completed_at:
        type: string
      created_at:
        type: string
      expired_at:
        type: string
      game_account_data:
        allOf:
        - $ref: '#/definitions/models.GameAccountData'
        description: Game Account Info (denormalized for transaction history)
      id:
        type: integer
      ip_address:
        type: string
      logs:
        items:
          $ref: '#/definitions/models.TransactionLog'
        type: array
      payment_fee:
        minimum: 0
        type: number
      payment_method:
        description: Payment
        enum:
        - gopay
        - ovo
        - dana
        - bca
        - mandiri
        - bni
        type: string
      payment_reference:
        type: string
      payment_status:
        enum:
        - pending
        - paid
        - failed
        - expired
        - refunded
        type: string
      payment_url:
        type: string
      processed_at:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: integer
      product_price:
        description: Pricing
        minimum: 0
        type: number
      status:
        description: Status & Processing
        enum:
        - pending
        - processing
        - completed
        - failed
        - cancelled
        type: string
      total_amount:
        minimum: 0
        type: number
      transaction_code:
        type: string
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relations
      user_agent:
        description: Metadata
        type: string
      user_id:
        type: integer
      voucher_discount:
        minimum: 0
        type: number
      voucher_usages:
        items:
          $ref: '#/definitions/models.VoucherUsage'
        type: array
      whatsapp:
        description: Contact
        type: string
    required:
    - game_account_data
    - payment_method
    - product_id
    - product_price
    - total_amount
    - user_id
    - whatsapp
    type: object
...
  utils.APIResponse:
    properties:
      data: {}
      errors:
        items:
          type: string
        type: array
      message:
        type: string
      pagination:
        $ref: '#/definitions/utils.Pagination'
      success:
        type: boolean
    type: object
  utils.Pagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@wawstore.com
    name: API Support
    url: https://wawstore.com/support
  description: Backend API for Waw Store - Game Topup Platform
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://wawstore.com/terms
  title: Waw Store Topup Game Online API
  version: "1.0"
paths:
  /admin/analytics/dashboard:
    get:
      description: Get dashboard analytics data for admin
      parameters:
      - default: month
        description: Analytics period
        enum:
        - today
        - week
        - month
        - year
        in: query
        name: period
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  properties:
                    pending_transactions:
                      type: integer
                    total_revenue:
                      type: number
                    total_transactions:
                      type: integer
                    total_users:
                      type: integer
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get dashboard analytics
      tags:
      - admin
...
  /admin/categories/{id}/upload-icon:
    post:
      consumes:
      - multipart/form-data
      description: Upload icon for a category (Admin only)
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Category icon file
        in: formData
        name: icon
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  properties:
                    icon_url:
                      type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Upload category icon
      tags:
      - admin
...
  /admin/products/{id}/upload-image:
    post:
      consumes:
      - multipart/form-data
      description: Upload image for a product (Admin only)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product image file
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  properties:
                    image_url:
                      type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Upload product image
      tags:
      - admin
  /admin/transactions:
    get:
      description: Get paginated list of all transactions for admin management
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by status
        enum:
        - pending
        - processing
        - completed
        - failed
        - cancelled
        in: query
        name: status
        type: string
      - description: Filter by user ID
        in: query
        name: user_id
        type: integer
      - description: Filter from date (YYYY-MM-DD)
        in: query
        name: date_from
        type: string
      - description: Filter to date (YYYY-MM-DD)
        in: query
        name: date_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Transaction'
                  type: array
                pagination:
                  $ref: '#/definitions/utils.Pagination'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get transactions for admin
      tags:
      - admin
  /admin/transactions/{id}/status:
    put:
      consumes:
      - application/json
      description: Update transaction status (Admin only)
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      - description: Status update request
        in: body
        name: request
        required: true
        schema:
          properties:
            message:
              type: string
            status:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Transaction'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Update transaction status
      tags:
      - admin
...
  /admin/vouchers:
    get:
      description: Get paginated list of vouchers for admin management
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by status
        enum:
        - active
        - inactive
        - expired
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Voucher'
                  type: array
                pagination:
                  $ref: '#/definitions/utils.Pagination'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get vouchers for admin
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: Create a new voucher (Admin only)
      parameters:
      - description: Voucher data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Voucher'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Voucher'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Create voucher
      tags:
      - admin
...
  /game-accounts:
    get:
      description: Get all game accounts for the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.GameAccount'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get user game accounts
      tags:
      - game-accounts
    post:
      consumes:
      - application/json
      description: Create a new game account for the current user
      parameters:
      - description: Game account data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.GameAccount'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.GameAccount'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Create game account
      tags:
      - game-accounts
  /game-accounts/{id}:
    delete:
      description: Delete a game account by ID
      parameters:
      - description: Game Account ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete game account
      tags:
      - game-accounts
    put:
      consumes:
      - application/json
      description: Update a game account by ID
      parameters:
      - description: Game Account ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated game account data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.GameAccount'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.GameAccount'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Update game account
      tags:
      - game-accounts
...
  /transactions:
    get:
      description: Get paginated list of current user's transactions
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by status
        enum:
        - pending
        - processing
        - completed
        - failed
        - cancelled
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Transaction'
                  type: array
                pagination:
                  $ref: '#/definitions/utils.Pagination'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get user transactions
      tags:
      - transactions
    post:
      consumes:
      - application/json
      description: Create a new topup transaction
      parameters:
      - description: Transaction data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Transaction'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Transaction'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Create transaction
      tags:
      - transactions
  /transactions/{id}:
    get:
      description: Get detailed transaction information by ID
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Transaction'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get transaction by ID
      tags:
      - transactions
  /transactions/{id}/cancel:
    post:
      description: Cancel a pending transaction
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Transaction'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Cancel transaction
      tags:
      - transactions
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"