# internal/handlers/placeholder_handlers.go
package handlers

import (
	"net/http"

	"github.com/gin-gonic/gin"

	"topup-backend/internal/services"
	"topup-backend/internal/utils"
)

// GameAccountHandler handles game account-related HTTP requests
type GameAccountHandler struct {
	services *services.ServiceContainer
}

// NewGameAccountHandler creates a new game account handler
func NewGameAccountHandler(services *services.ServiceContainer) *GameAccountHandler {
	return &GameAccountHandler{
		services: services,
	}
}

// GetUserGameAccounts godoc
// @Summary Get user game accounts
// @Description Get all game accounts for the current user
// @Tags game-accounts
// @Security BearerAuth
// @Produce json
// @Success 200 {object} utils.APIResponse{data=[]models.GameAccount}
// @Failure 401 {object} utils.APIResponse
// @Failure 500 {object} utils.APIResponse
// @Router /game-accounts [get]
func (h *GameAccountHandler) GetUserGameAccounts(c *gin.Context) {
	utils.ErrorResponse(c, http.StatusNotImplemented, "Not implemented yet")
}

// CreateGameAccount godoc
// @Summary Create game account
// @Description Create a new game account for the current user
// @Tags game-accounts
// @Security BearerAuth
// @Accept json
// @Produce json
// @Param request body models.GameAccount true "Game account data"
// @Success 201 {object} utils.APIResponse{data=models.GameAccount}
// @Failure 400 {object} utils.APIResponse
// @Failure 401 {object} utils.APIResponse
// @Failure 500 {object} utils.APIResponse
// @Router /game-accounts [post]
func (h *GameAccountHandler) CreateGameAccount(c *gin.Context) {
	utils.ErrorResponse(c, http.StatusNotImplemented, "Not implemented yet")
}

// UpdateGameAccount godoc
// @Summary Update game account
// @Description Update a game account by ID
// @Tags game-accounts
// @Security BearerAuth
// @Accept json
// @Produce json
// @Param id path int true "Game Account ID"
// @Param request body models.GameAccount true "Updated game account data"
// @Success 200 {object} utils.APIResponse{data=models.GameAccount}
// @Failure 400 {object} utils.APIResponse
// @Failure 401 {object} utils.APIResponse
// @Failure 404 {object} utils.APIResponse
// @Failure 500 {object} utils.APIResponse
// @Router /game-accounts/{id} [put]
func (h *GameAccountHandler) UpdateGameAccount(c *gin.Context) {
	utils.ErrorResponse(c, http.StatusNotImplemented, "Not implemented yet")
}

// DeleteGameAccount godoc
// @Summary Delete game account
// @Description Delete a game account by ID
// @Tags game-accounts
// @Security BearerAuth
// @Produce json
// @Param id path int true "Game Account ID"
// @Success 200 {object} utils.APIResponse
// @Failure 401 {object} utils.APIResponse
// @Failure 404 {object} utils.APIResponse
// @Failure 500 {object} utils.APIResponse
// @Router /game-accounts/{id} [delete]
func (h *GameAccountHandler) DeleteGameAccount(c *gin.Context) {
	utils.ErrorResponse(c, http.StatusNotImplemented, "Not implemented yet")
}

// VoucherHandler handles voucher-related HTTP requests
type VoucherHandler struct {
	services *services.ServiceContainer
}

// NewVoucherHandler creates a new voucher handler
func NewVoucherHandler(services *services.ServiceContainer) *VoucherHandler {
	return &VoucherHandler{
		services: services,
	}
}

// ValidateVoucher godoc
// @Summary Validate voucher
// @Description Validate a voucher code and return discount information
// @Tags vouchers
// @Accept json
// @Produce json
// @Param request body object{code=string,product_id=int,total_amount=number} true "Voucher validation request"
// @Success 200 {object} utils.APIResponse{data=object{valid=boolean,discount_amount=number,message=string}}
// @Failure 400 {object} utils.APIResponse
// @Failure 404 {object} utils.APIResponse
// @Failure 500 {object} utils.APIResponse
// @Router /vouchers/validate [post]
func (h *VoucherHandler) ValidateVoucher(c *gin.Context) {
	utils.ErrorResponse(c, http.StatusNotImplemented, "Not implemented yet")
}

// GetVouchersForAdmin godoc
// @Summary Get vouchers for admin
// @Description Get paginated list of vouchers for admin management
// @Tags admin
// @Security BearerAuth
// @Produce json
// @Param page query int false "Page number" default(1)
// @Param limit query int false "Items per page" default(10)
// @Param status query string false "Filter by status" Enums(active, inactive, expired)
// @Success 200 {object} utils.APIResponse{data=[]models.Voucher,pagination=utils.Pagination}
// @Failure 401 {object} utils.APIResponse
// @Failure 500 {object} utils.APIResponse
// @Router /admin/vouchers [get]
func (h *VoucherHandler) GetVouchersForAdmin(c *gin.Context) {
	utils.ErrorResponse(c, http.StatusNotImplemented, "Not implemented yet")
}

// CreateVoucher godoc
// @Summary Create voucher
// @Description Create a new voucher (Admin only)
// @Tags admin
// @Security BearerAuth
// @Accept json
// @Produce json
// @Param request body models.Voucher true "Voucher data"
// @Success 201 {object} utils.APIResponse{data=models.Voucher}
// @Failure 400 {object} utils.APIResponse
// @Failure 401 {object} utils.APIResponse
// @Failure 403 {object} utils.APIResponse
// @Failure 500 {object} utils.APIResponse
// @Router /admin/vouchers [post]
func (h *VoucherHandler) CreateVoucher(c *gin.Context) {
	utils.ErrorResponse(c, http.StatusNotImplemented, "Not implemented yet")
}

// GetVoucherByID gets voucher by ID
func (h *VoucherHandler) GetVoucherByID(c *gin.Context) {
	utils.ErrorResponse(c, http.StatusNotImplemented, "Not implemented yet")
}

// UpdateVoucher updates a voucher
func (h *VoucherHandler) UpdateVoucher(c *gin.Context) {
	utils.ErrorResponse(c, http.StatusNotImplemented, "Not implemented yet")
}

// DeleteVoucher deletes a voucher
func (h *VoucherHandler) DeleteVoucher(c *gin.Context) {
	utils.ErrorResponse(c, http.StatusNotImplemented, "Not implemented yet")
}

// GetVoucherUsageStats gets voucher usage statistics
func (h *VoucherHandler) GetVoucherUsageStats(c *gin.Context) {
	utils.ErrorResponse(c, http.StatusNotImplemented, "Not implemented yet")
}

// TransactionHandler handles transaction-related HTTP requests
type TransactionHandler struct {
	services *services.ServiceContainer
}

// NewTransactionHandler creates a new transaction handler
func NewTransactionHandler(services *services.ServiceContainer) *TransactionHandler {
	return &TransactionHandler{
		services: services,
	}
}

// CreateTransaction godoc
// @Summary Create transaction
// @Description Create a new topup transaction
// @Tags transactions
// @Security BearerAuth
// @Accept json
// @Produce json
// @Param request body models.Transaction true "Transaction data"
// @Success 201 {object} utils.APIResponse{data=models.Transaction}
// @Failure 400 {object} utils.APIResponse
// @Failure 401 {object} utils.APIResponse
// @Failure 500 {object} utils.APIResponse
// @Router /transactions [post]
func (h *TransactionHandler) CreateTransaction(c *gin.Context) {
	utils.ErrorResponse(c, http.StatusNotImplemented, "Not implemented yet")
}

// GetUserTransactions godoc
// @Summary Get user transactions
// @Description Get paginated list of current user's transactions
// @Tags transactions
// @Security BearerAuth
// @Produce json
// @Param page query int false "Page number" default(1)
// @Param limit query int false "Items per page" default(10)
// @Param status query string false "Filter by status" Enums(pending, processing, completed, failed, cancelled)
// @Success 200 {object} utils.APIResponse{data=[]models.Transaction,pagination=utils.Pagination}
// @Failure 401 {object} utils.APIResponse
// @Failure 500 {object} utils.APIResponse
// @Router /transactions [get]
func (h *TransactionHandler) GetUserTransactions(c *gin.Context) {
	utils.ErrorResponse(c, http.StatusNotImplemented, "Not implemented yet")
}

// GetTransactionByID godoc
// @Summary Get transaction by ID
// @Description Get detailed transaction information by ID
// @Tags transactions
// @Security BearerAuth
// @Produce json
// @Param id path int true "Transaction ID"
// @Success 200 {object} utils.APIResponse{data=models.Transaction}
// @Failure 401 {object} utils.APIResponse
// @Failure 404 {object} utils.APIResponse
// @Failure 500 {object} utils.APIResponse
// @Router /transactions/{id} [get]
func (h *TransactionHandler) GetTransactionByID(c *gin.Context) {
	utils.ErrorResponse(c, http.StatusNotImplemented, "Not implemented yet")
}

// CancelTransaction godoc
// @Summary Cancel transaction
// @Description Cancel a pending transaction
// @Tags transactions
// @Security BearerAuth
// @Produce json
// @Param id path int true "Transaction ID"
// @Success 200 {object} utils.APIResponse{data=models.Transaction}
// @Failure 400 {object} utils.APIResponse
// @Failure 401 {object} utils.APIResponse
// @Failure 404 {object} utils.APIResponse
// @Failure 500 {object} utils.APIResponse
// @Router /transactions/{id}/cancel [post]
func (h *TransactionHandler) CancelTransaction(c *gin.Context) {
	utils.ErrorResponse(c, http.StatusNotImplemented, "Not implemented yet")
}

// GetTransactionsForAdmin godoc
// @Summary Get transactions for admin
// @Description Get paginated list of all transactions for admin management
// @Tags admin
// @Security BearerAuth
// @Produce json
// @Param page query int false "Page number" default(1)
// @Param limit query int false "Items per page" default(10)
// @Param status query string false "Filter by status" Enums(pending, processing, completed, failed, cancelled)
// @Param user_id query int false "Filter by user ID"
// @Param date_from query string false "Filter from date (YYYY-MM-DD)"
// @Param date_to query string false "Filter to date (YYYY-MM-DD)"
// @Success 200 {object} utils.APIResponse{data=[]models.Transaction,pagination=utils.Pagination}
// @Failure 401 {object} utils.APIResponse
// @Failure 500 {object} utils.APIResponse
// @Router /admin/transactions [get]
func (h *TransactionHandler) GetTransactionsForAdmin(c *gin.Context) {
	utils.ErrorResponse(c, http.StatusNotImplemented, "Not implemented yet")
}

// GetTransactionByIDForAdmin gets transaction by ID for admin
func (h *TransactionHandler) GetTransactionByIDForAdmin(c *gin.Context) {
	utils.ErrorResponse(c, http.StatusNotImplemented, "Not implemented yet")
}

// UpdateTransactionStatus godoc
// @Summary Update transaction status
// @Description Update transaction status (Admin only)
// @Tags admin
// @Security BearerAuth
// @Accept json
// @Produce json
// @Param id path int true "Transaction ID"
// @Param request body object{status=string,message=string} true "Status update request"
// @Success 200 {object} utils.APIResponse{data=models.Transaction}
// @Failure 400 {object} utils.APIResponse
// @Failure 401 {object} utils.APIResponse
// @Failure 404 {object} utils.APIResponse
// @Failure 500 {object} utils.APIResponse
// @Router /admin/transactions/{id}/status [put]
func (h *TransactionHandler) UpdateTransactionStatus(c *gin.Context) {
	utils.ErrorResponse(c, http.StatusNotImplemented, "Not implemented yet")
}

// ExportTransactions exports transactions
func (h *TransactionHandler) ExportTransactions(c *gin.Context) {
	utils.ErrorResponse(c, http.StatusNotImplemented, "Not implemented yet")
}

// FileUploadHandler handles file upload-related HTTP requests
type FileUploadHandler struct {
	services *services.ServiceContainer
}

// NewFileUploadHandler creates a new file upload handler
func NewFileUploadHandler(services *services.ServiceContainer) *FileUploadHandler {
	return &FileUploadHandler{
		services: services,
	}
}

// UploadProductImage godoc
// @Summary Upload product image
// @Description Upload image for a product (Admin only)
// @Tags admin
// @Security BearerAuth
// @Accept multipart/form-data
// @Produce json
// @Param id path int true "Product ID"
// @Param image formData file true "Product image file"
// @Success 200 {object} utils.APIResponse{data=object{image_url=string}}
// @Failure 400 {object} utils.APIResponse
// @Failure 401 {object} utils.APIResponse
// @Failure 403 {object} utils.APIResponse
// @Failure 404 {object} utils.APIResponse
// @Failure 500 {object} utils.APIResponse
// @Router /admin/products/{id}/upload-image [post]
func (h *FileUploadHandler) UploadProductImage(c *gin.Context) {
	utils.ErrorResponse(c, http.StatusNotImplemented, "Not implemented yet")
}

// UploadCategoryIcon godoc
// @Summary Upload category icon
// @Description Upload icon for a category (Admin only)
// @Tags admin
// @Security BearerAuth
// @Accept multipart/form-data
// @Produce json
// @Param id path int true "Category ID"
// @Param icon formData file true "Category icon file"
// @Success 200 {object} utils.APIResponse{data=object{icon_url=string}}
// @Failure 400 {object} utils.APIResponse
// @Failure 401 {object} utils.APIResponse
// @Failure 403 {object} utils.APIResponse
// @Failure 404 {object} utils.APIResponse
// @Failure 500 {object} utils.APIResponse
// @Router /admin/categories/{id}/upload-icon [post]
func (h *FileUploadHandler) UploadCategoryIcon(c *gin.Context) {
	utils.ErrorResponse(c, http.StatusNotImplemented, "Not implemented yet")
}

// AnalyticsHandler handles analytics-related HTTP requests
type AnalyticsHandler struct {
	services *services.ServiceContainer
}

// NewAnalyticsHandler creates a new analytics handler
func NewAnalyticsHandler(services *services.ServiceContainer) *AnalyticsHandler {
	return &AnalyticsHandler{
		services: services,
	}
}

// GetDashboardAnalytics godoc
// @Summary Get dashboard analytics
// @Description Get dashboard analytics data for admin
// @Tags admin
// @Security BearerAuth
// @Produce json
// @Param period query string false "Analytics period" Enums(today, week, month, year) default(month)
// @Success 200 {object} utils.APIResponse{data=object{total_transactions=int,total_revenue=number,total_users=int,pending_transactions=int}}
// @Failure 401 {object} utils.APIResponse
// @Failure 500 {object} utils.APIResponse
// @Router /admin/analytics/dashboard [get]
func (h *AnalyticsHandler) GetDashboardAnalytics(c *gin.Context) {
	utils.ErrorResponse(c, http.StatusNotImplemented, "Not implemented yet")
}